pyproject.toml
[tool.poetry]
name = "first"
version = "0.1.0"
description = ""
authors = ["nurj0hn <nurjantunteev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aioredis = "2.0.1"
aiosmtplib = "2.0.2"
alembic = "1.11.1"
anyio = "3.7.0"
async-timeout = "4.0.3"
asyncpg = "0.27.0"
bcrypt = "4.0.1"
blinker = "1.6.2"
certifi = "2023.7.22"
cfgv = "3.3.1"
click = "8.1.3"
distlib = "0.3.6"
dnspython = "2.3.0"
ecdsa = "0.18.0"
email-validator = "2.0.0.post2"
envparse = "0.2.0"
exceptiongroup = "1.1.3"
fakeredis = "2.18.1"
fastapi = "0.98.0"
fastapi-mail = "0.4.1"
filelock = "3.12.2"
flask = "2.3.3"
flask-admin = "1.6.1"
greenlet = "2.0.2"
h11 = "0.14.0"
httpcore = "0.17.3"
httpx = "0.24.1"
identify = "2.5.24"
idna = "3.4"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
mako = "1.2.4"
markupsafe = "2.1.3"
nodeenv = "1.8.0"
passlib = "1.7.4"
platformdirs = "3.8.0"
pre-commit = "3.3.3"
psycopg2 = "2.9.6"
psycopg2-binary = "2.9.6"
pyasn1 = "0.5.0"
pydantic = "1.10.9"
python-decouple = "3.8"
python-jose = "3.3.0"
python-multipart = "0.0.6"
pyyaml = "6.0"
redis = "5.0.0"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.0"
sortedcontainers = "2.4.0"
sqlalchemy = "2.0.17"
starlette = "0.27.0"
typing-extensions = "4.7.0"
uvicorn = "0.22.0"
virtualenv = "20.24.1"
werkzeug = "2.3.7"
wtforms = "3.0.1"
flask-wtf = "1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
